stream_settings_ui: Extract stream_settings_components module.
